requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
jinja2>=3.1.2
python-dotenv>=1.0.0
pydantic>=2.4.0
numpy>=1.24.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.33.0
scikit-learn>=1.3.0
pandas>=2.0.0
kagglehub>=0.3.5
tqdm>=4.66.1
sentencepiece>=0.2.0
